move files for injected.bundle.js into separate folder clean up filenames
