+++ src/pages/Popup/Popup.tsx
@@ -16,6 +16,7 @@ export const Popup = (): JSX.Element => {
   }, [pbjsNamespace]);
 
   useEffect(() => {
+    // console.log('popup POPUP_LOADED', );  
     sendChromeTabsMessage(POPUP_LOADED, {});
   }, []);
