+++ src/pages/Popup/components/config/FirstPartyDataComponent.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { IPrebidConfig } from '../../../Content/scripts/prebid';
+import { IPrebidConfig } from '../../../Injected/prebid';
 import Typography from '@mui/material/Typography';
 import Avatar from '@mui/material/Avatar';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
