+++ src/pages/Shared/components/adUnits/grid/tiles/AdServerTile.tsx
@@ -2,9 +2,9 @@ import React, { useContext, useState, useEffect } from 'react';
 import Stack from '@mui/material/Stack';
 import Chip from '@mui/material/Chip';
 import Typography from '@mui/material/Typography';
-import { IPrebidAdUnit } from '../../../../../Content/scripts/prebid';
+import { IPrebidAdUnit } from '../../../../../Injected/prebid';
 import Box from '@mui/material/Box';
-import { IGoogleAdManagerSlot } from '../../../../../Content/scripts/googleAdManager';
+import { IGoogleAdManagerSlot } from '../../../../../Injected/googleAdManager';
 import JSONViewerComponent from '../../../JSONViewerComponent';
 import Grid from '@mui/material/Grid';
 import AppStateContext from '../../../../contexts/appStateContext';
