+++ src/pages/Shared/contexts/fetchEvents.ts
@@ -1,6 +1,6 @@
 import { getTabId, sendChromeTabsMessage } from '../../Shared/utils';
 import { DOWNLOAD_FAILED } from '../constants';
-import { ITabInfos } from '../../Background/background';
+import { ITabInfos } from '../../Background';
 
 const safelyConstructURL = (url: string) => {
   try {
