+++ webpack.config.js
@@ -29,12 +29,12 @@ var options = {
     newtab: path.join(__dirname, 'src', 'pages', 'Newtab', 'index.jsx'),
     options: path.join(__dirname, 'src', 'pages', 'Options', 'index.jsx'),
     popup: path.join(__dirname, 'src', 'pages', 'Popup', 'index.tsx'),
-    background: path.join(__dirname, 'src', 'pages', 'Background', 'background.ts'),
+    background: path.join(__dirname, 'src', 'pages', 'Background', 'index.ts'),
     contentScript: path.join(__dirname, 'src', 'pages', 'Content', 'index.ts'),
     devtools: path.join(__dirname, 'src', 'pages', 'Devtools', 'index.ts'),
     panel: path.join(__dirname, 'src', 'pages', 'Panel', 'index.tsx'),
-    injected: path.join(__dirname, 'src', 'pages', 'Content', 'scripts', 'injected.tsx'),
-    openDfpConsole: path.join(__dirname, 'src', 'pages', 'Content', 'scripts', 'openDfpConsole.ts'),
+    injected: path.join(__dirname, 'src', 'pages', 'Injected', 'index.tsx'),
+    openDfpConsole: path.join(__dirname, 'src', 'pages', 'Injected', 'openDfpConsole.ts'),
   },
   chromeExtensionBoilerplate: {
     notHotReload: ['background', 'background.bundle.js', 'contentScript', 'contentScript.bundle.js', 'devtools', 'injected', 'openDfpConsole'],
